Resolved -Wall warning: cast to long for sizeof/%ld so they jive on 32-bit systems.

+++ src/cdnaAli/g2cCheck/g2cCheck.c
@@ -126,7 +126,7 @@ while (fgets(lineBuf, sizeof(lineBuf), f) != NULL)
     if (lineLen >= sizeof(lineBuf) - 1)
         {
         errAbort("%s\nLine %d of %s too long, can only handle %ld chars\n",
-            lineBuf, lineCount, fileName, sizeof(lineBuf)-1);
+            lineBuf, lineCount, fileName, (long)sizeof(lineBuf)-1);
         }
     wordCount = chopString(lineBuf, whiteSpaceChopper, words, ArraySize(words));
     if (wordCount > 0)
